Now with two queues
Untested- will try out later.  The idea is that we have a Q in the main thread which can call methods on the view object, and one in the worker thread.  There are two separate set timeout loops- one, collects changes in views and shoves them into the the worker queue.  The other gets runs dmp and deals with networking.